HIF support for Idol: flesh out interface
