+++ humility-cmd/Cargo.toml
@@ -9,6 +9,7 @@ clap = "2.33.0"
 structopt = "0.3"
 anyhow = { version = "1.0.44", features = ["backtrace"] }
 hif = { git = "https://github.com/oxidecomputer/hif" }
+idol = {git = "https://github.com/oxidecomputer/idolatry.git"}
 indexmap = { version = "1.7", features = ["serde-1"] }
 humility_load_derive = {path = "../load_derive"}
 postcard = "0.7.0"
